requirements.txt
websockets>=12.0
websocket-client>=1.6
requests>=2.31
pandas>=2.0
python-dotenv>=1.0
